edit crop image + edit image path + edit remove image

+++ app/Http/Controllers/PostsController.php
@@ -87,20 +87,16 @@ class PostsController extends Controller
                     $front_image_title[$file->getClientOriginalExtension() . '_' . $key] = $file_name;
                     $file->move(public_path("/posts/front_image/" . date('F_Y')), $file_name);
 
-                    /* crop image */
+                    /* resize image */
                     $path = $_SERVER['DOCUMENT_ROOT'] . "/posts/front_image/" . date('F_Y') . '/' . $file_name;
                     $image = Image::make($path);
 
-                    /* resize & crop image */
                     list($width, $height) = getimagesize($path);
-                    $ratio = 16 / 9;
+                    $new_width = 1200;
+                    $ratio = $width / $height;
 
-                    foreach($size as $image_size) {
-                        if($width < $image_size['width'] && $width > $image_size['next_width']) {
-                            $image->fit($image_size['next_width'], intval($image_size['next_width'] / $ratio));
-                        } elseif($width == $image_size['width']){
-                            $image->fit($image_size['width'], intval($image_size['width'] / $ratio));
-                        }
+                    if($width > $new_width) {
+                        $image->resize($new_width, intval($new_width / $ratio));
                     }
 
                     /* save new image */
@@ -129,20 +125,16 @@ class PostsController extends Controller
                         $body_image_title[$file->getClientOriginalExtension() . '_' . $key] = $file_name;
                         $file->move(public_path("/posts/body_image/" . date('F_Y')), $file_name);
 
-                        /* crop image */
+                        /* resize image */
                         $path = $_SERVER['DOCUMENT_ROOT'] . "/posts/body_image/" . date('F_Y') . '/' . $file_name;
                         $image = Image::make($path);
 
-                        /* resize & crop image */
                         list($width, $height) = getimagesize($path);
-                        $ratio = 16 / 9;
+                        $new_width = 1200;
+                        $ratio = $width / $height;
 
-                        foreach($size as $image_size) {
-                            if($width < $image_size['width'] && $width > $image_size['next_width']) {
-                                $image->fit($image_size['next_width'], intval($image_size['next_width'] / $ratio));
-                            } elseif($width == $image_size['width']){
-                                $image->fit($image_size['width'], intval($image_size['width'] / $ratio));
-                            }
+                        if($width > $new_width) {
+                            $image->resize($new_width, intval($new_width / $ratio));
                         }
 
                         /* save new image */
@@ -227,15 +219,6 @@ class PostsController extends Controller
 
             $front_image = $request->file('front_image');
             $body_image = $request->file('body_image');
-            $size = [
-                ['width' => '9999', 'next_width' => '3200'],
-                ['width' => '3200', 'next_width' => '2880'], ['width' => '2880', 'next_width' => '2560'],
-                ['width' => '2560', 'next_width' => '2048'], ['width' => '2048', 'next_width' => '1920'],
-                ['width' => '1920', 'next_width' => '1600'], ['width' => '1600', 'next_width' => '1366'],
-                ['width' => '1366', 'next_width' => '1280'], ['width' => '1280', 'next_width' => '1024'],
-                ['width' => '1024', 'next_width' => '960'], ['width' => '960', 'next_width' => '864'],
-                ['width' => '864', 'next_width' => '720'], ['width' => '720', 'next_width' => '640'],
-            ];
 
             if($request->hasFile('front_image'))
             {
@@ -244,20 +227,16 @@ class PostsController extends Controller
                     $front_image_title[$file->getClientOriginalExtension() . '_' . $key] = $file_name;
                     $file->move(public_path("/posts/front_image/" . date('F_Y')), $file_name);
 
-                    /* crop image */
+                    /* resize image */
                     $path = $_SERVER['DOCUMENT_ROOT'] . "/posts/front_image/" . date('F_Y') . '/' . $file_name;
                     $image = Image::make($path);
 
-                    /* resize & crop image */
                     list($width, $height) = getimagesize($path);
-                    $ratio = 16 / 9;
+                    $new_width = 1200;
+                    $ratio = $width / $height;
 
-                    foreach($size as $image_size) {
-                        if($width < $image_size['width'] && $width > $image_size['next_width']) {
-                            $image->fit($image_size['next_width'], intval($image_size['next_width'] / $ratio));
-                        } elseif($width == $image_size['width']){
-                            $image->fit($image_size['width'], intval($image_size['width'] / $ratio));
-                        }
+                    if($width > $new_width) {
+                        $image->resize($new_width, intval($new_width / $ratio));
                     }
 
                     /* save new image */
@@ -286,20 +265,16 @@ class PostsController extends Controller
                         $body_image_title[$file->getClientOriginalExtension() . '_' . $key] = $file_name;
                         $file->move(public_path("/posts/body_image/" . date('F_Y')), $file_name);
 
-                        /* crop image */
+                        /* resize image */
                         $path = $_SERVER['DOCUMENT_ROOT'] . "/posts/body_image/" . date('F_Y') . '/' . $file_name;
                         $image = Image::make($path);
 
-                        /* resize & crop image */
                         list($width, $height) = getimagesize($path);
-                        $ratio = 16 / 9;
+                        $new_width = 1200;
+                        $ratio = $width / $height;
 
-                        foreach($size as $image_size) {
-                            if($width < $image_size['width'] && $width > $image_size['next_width']) {
-                                $image->fit($image_size['next_width'], intval($image_size['next_width'] / $ratio));
-                            } elseif($width == $image_size['width']){
-                                $image->fit($image_size['width'], intval($image_size['width'] / $ratio));
-                            }
+                        if($width > $new_width) {
+                            $image->resize($new_width, intval($new_width / $ratio));
                         }
 
                         /* save new image */
@@ -341,14 +316,19 @@ class PostsController extends Controller
             foreach (json_decode($item['front_image'], true) as $image_path) {
                 if(!empty($item['front_image'])) {
                     $front_image_path = $_SERVER['DOCUMENT_ROOT'] . "/posts/front_image/" . date('F_Y') . '/' . $image_path;
-                    unlink($front_image_path);
+                    if(file_exists($front_image_path)) {
+                        unlink($front_image_path);
+                    }
+
                 }
             }
 
             foreach (json_decode($item['body_image'], true) as $image_path) {
                 if(!empty($item['body_image'])) {
                     $body_image_path = $_SERVER['DOCUMENT_ROOT'] . "/posts/body_image/" . date('F_Y') . '/' . $image_path;
-                    unlink($body_image_path);
+                    if(file_exists($body_image_path)) {
+                        unlink($body_image_path);
+                    }
                 }
             }
             $posts->delete();
